refactoring alpine data chunks into their own files
